Refactor transaction history to own component
