+++ client/src/pages/singleGiftCard.jsx
@@ -1,14 +1,16 @@
 import "../components/GiftCardListItem.css";
-import { useState, useContext, useEffect } from "react";
-import { useLocation } from "react-router-dom";
-import { useNavigate } from "react-router-dom";
+import { useState, useContext } from "react";
+import { useLocation, useNavigate, useParams } from "react-router-dom";
 import CodeView from "../components/CodeView";
 import Button from "../components/Button";
 import CreditCard from "../components/CreditCard";
 import axios from "axios";
 import LoggedInUser from "../context/AuthContext";
+import CardTransactions from "../components/CardTransactions";
 
 const SingleGiftCard = (props) => {
+  let params = useParams();
+  
   const { state } = useLocation();
   const {
     name,
@@ -25,7 +27,6 @@ const SingleGiftCard = (props) => {
   const [card, setCard] = useState(false);
   const [text, setText] = useState("Place Order");
   const [qrCode, setQrCode] = useState(false);
-  const [transactions, setTransactions] = useState([]);
   const context = useContext(LoggedInUser);
   let navigate = useNavigate();
 
@@ -55,23 +56,6 @@ const SingleGiftCard = (props) => {
 
       .catch((err) => console.log(err.message));
   };
-
-  useEffect(() => {
-    axios
-      .get(`/cards/${state.gift_card_id}/transactions`)
-      .then((response) => setTransactions(response.data));
-  }, [state.gift_card_id]);
-
-  // map to render transaction table rows
-  const transactionTable = transactions.map((data) => {
-    const formattedDate = new Date(data.created_at);
-    return (
-      <tr>
-        <td>{formatter.format(data.amount / 100)}</td>
-        <td>{formattedDate.toLocaleString()}</td>
-      </tr>
-    );
-  });
 
   return (
     <div className="single-card-content">
@@ -130,18 +114,7 @@ const SingleGiftCard = (props) => {
           onPay={onPay}
         />
       )}
-      <div className="transaction-list">
-        <h2>Transaction History</h2>
-        <table>
-          <thead>
-            <tr>
-              <th>Amount</th>
-              <th>Time</th>
-            </tr>
-          </thead>
-          <tbody>{transactionTable}</tbody>
-        </table>
-      </div>
+      <CardTransactions key={params.id} cardID={params.id} />
     </div>
   );
 };
